Various fiddlings
Including commenting some things out, and also changing the name
of db.just_pulled.get to be get_from_disk, to make it clearer that
it is replacing the previous (internal) value

+++ muddled/db.py
@@ -393,12 +393,12 @@ class Database(object):
         # If the subdomain just pulled stuff (as it will have done if it
         # checked anything out, because that counts), then we need to add
         # it to *our* set of just_pulled stuff
-        print 'JUST PULLED MERGING IN %s'%other_db.just_pulled.file_name[34:]
-        if self.just_pulled.labels or other_db.just_pulled.labels:
-            print 'JUST PULLED FOR %s was     %s'%(self.just_pulled.file_name[34:], ', '.join(map(str, self.just_pulled.labels)))
+        ##print 'JUST PULLED MERGING IN %s'%other_db.just_pulled.file_name[34:]
+        ##if self.just_pulled.labels or other_db.just_pulled.labels:
+        ##    print 'JUST PULLED FOR %s was     %s'%(self.just_pulled.file_name[34:], ', '.join(map(str, self.just_pulled.labels)))
         self.just_pulled.labels.update(other_db.just_pulled.labels)
-        if self.just_pulled.labels or other_db.just_pulled.labels:
-            print 'JUST PULLED FOR %s becomes %s'%(self.just_pulled.file_name[34:], ', '.join(map(str, self.just_pulled.labels)))
+        ##if self.just_pulled.labels or other_db.just_pulled.labels:
+        ##    print 'JUST PULLED FOR %s becomes %s'%(self.just_pulled.file_name[34:], ', '.join(map(str, self.just_pulled.labels)))
 
     def _merge_subdomain_upstreams(self, other_domain_name, other_db):
         """Merge things from the subdomain that contain upstream repositories.
@@ -1725,13 +1725,13 @@ class JustPulledFile(object):
         self.file_name = file_name
         self.labels = set()
 
-    def get(self):
+    def get_from_disk(self):
         """Retrieve the contents of the _just_pulled file as a list of labels.
 
         First clears the local memory, then reads the labels in the _just_pulled
         file into local memory, then returns that set as a sorted list.
         """
-        print 'XXX %s CLEAR AND GET JUST PULLED'%self.file_name[34:]
+        ##print 'XXX %s CLEAR AND GET JUST PULLED'%self.file_name[34:]
         self.labels.clear()
         try:
             line_no = 0
@@ -1758,11 +1758,14 @@ class JustPulledFile(object):
 
     def clear(self):
         """Clear the contents of the _just_pulled file, and our local memory.
+
+        If the _just_pulled files does not exist, does nothing
         """
-        print 'XXX %s CLEAR JUST PULLED'%self.file_name[34:]
+        ##print 'XXX %s CLEAR JUST PULLED'%self.file_name[34:]
         self.labels.clear()
-        with open(self.file_name, 'w') as fd:
-            pass
+        if os.path.exists(self.file_name):
+            with open(self.file_name, 'w') as fd:
+                pass
 
     def add(self, label):
         """Add the label to our local memory.
@@ -1770,7 +1773,7 @@ class JustPulledFile(object):
         The label is not added to the _just_pulled file until commit() is
         called.
         """
-        print 'XXX %s ADD JUST PULLED %s'%(self.file_name[34:], label)
+        ##print 'XXX %s ADD JUST PULLED %s'%(self.file_name[34:], label)
         self.labels.add(label.copy_with_tag(utils.LabelTag.CheckedOut))
 
     def commit(self):
@@ -1780,10 +1783,10 @@ class JustPulledFile(object):
 
         Leaves the local memory intact after writing (it does not clear it).
         """
-        print 'XXX %s COMMIT JUST PULLED'%self.file_name[34:]
+        ##print 'XXX %s COMMIT JUST PULLED'%self.file_name[34:]
         with open(self.file_name, 'w') as fd:
             for label in sorted(self.labels):
-                print 'XXX %s'%label
+                ##print 'XXX %s'%label
                 fd.write('%s\n'%label)
 
 # End file
